Fix obvious issues with registration, looking pretty good
