+++ app/Http/Controllers/Auth/LoginController.php
@@ -2,13 +2,11 @@
 
 namespace App\Http\Controllers\Auth;
 
-use App\Http\Controllers\Controller;
 use App\Models\Data\User;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
 use Illuminate\Http\Request;
-use Lcobucci\JWT\Builder;
 
-class LoginController extends Controller
+class LoginController extends AuthController
 {
     use AuthenticatesUsers {
         logout as performLogout;
@@ -47,24 +45,5 @@ class LoginController extends Controller
         $request->session()->put('jwt', $token);
 
         return response()->json($user);
-    }
-
-    /**
-     * Build the JWT used for authenticating API requests
-     *
-     * @param Request $request
-     * @return \Lcobucci\JWT\Token
-     */
-    protected function generateToken(Request $request)
-    {
-        return (new Builder())->setIssuer(config('app.url'))
-            ->setAudience(config('app.url'))
-            //->setId('4f1g23a12aa', true)
-            ->setIssuedAt(time())
-            ->setNotBefore(time() + 60)
-            ->setExpiration(time() + 3600)
-            ->set('uid', 1)
-            ->setSubject($request->input($this->username()))
-            ->getToken();
     }
 }
